Refactor validation types and enhance translation service
- Updated the validation types to improve type safety and consistency across the application.
- Refactored the IDSTranslationService to utilize new validation interfaces, streamlining translation processes.
- Enhanced the HTML report generation logic to ensure accurate rendering of validation results with updated entity properties.
